Add 5 initial test cases

+++ src/main/java/com/github/josephnun3z/pricesservice/boundary/RecommendedPriceDto.java
@@ -3,8 +3,8 @@ package com.github.josephnun3z.pricesservice.boundary;
 public record RecommendedPriceDto(
         Long productId,
         Integer brandId,
-        Long priceListId,
-        Float recommendedPrice,
+        Integer priceListId,
+        Double recommendedPrice,
         String currency
 ) {
 }
